fix swagger and user-routes
