+++ routes/api/auth-routes.js
@@ -2,7 +2,7 @@ const express = require("express");
 
 const ctrl = require("../../controllers/auth-controllers");
 
-const {authenticate, upload} = require("../../middlewares");
+const {authenticate} = require("../../middlewares");
 
 const { validateBody } = require("../../utils");
 
@@ -14,11 +14,7 @@ router.post("/register", validateBody(schemas.registerSchema), ctrl.register);
 
 router.post("/login", validateBody(schemas.loginSchema), ctrl.login);
 
-router.get("/current", authenticate, ctrl.getCurrent);
-
 router.post("/logout",  authenticate, ctrl.logout );
-
-router.patch("/avatars", authenticate, upload.single("avatar"), ctrl.updateAvatar);
 
 
 module.exports = router;
